longitude/Parking_excercise/parkinglot/requirements.txt
click==6.7
dj-database-url==0.4.2
Django==1.10.6
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-extensions==1.7.8
django-filter==1.0.2
django-tastypie==0.13.3
djangorestframework==3.6.2
Flask==0.12
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.0
psycopg2==2.7.1
PyMySQL==0.7.10
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
Werkzeug==0.12.1
